refactor: portfolio component for vanilla react
