+++ src/components/Card.tsx
@@ -1,6 +1,5 @@
-import * as React from "react"
-
-import { cn } from "@/lib/utils"
+import * as React from "react";
+import { cn } from "../utils";
 
 const Card = React.forwardRef<
   HTMLDivElement,
