+++ src/components/Portfolio.tsx
@@ -2,27 +2,26 @@
  * This code was generated by v0 by Vercel.
  * @see https://v0.dev/t/v6RyIF8aBrq
  */
-import Link from "next/link"
-import { CardContent, Card } from "@/components/ui/card"
+import { Card, CardContent } from "./Card"
 
 export function Portfolio() {
   return (
     <div className="flex flex-col min-h-screen">
       <header className="flex h-14 items-center justify-between px-4 md:px-6">
-        <Link className="flex items-center gap-2" href="#">
+        <a className="flex items-center gap-2" href="#">
           <MountainIcon className="h-6 w-6" />
           <span className="sr-only">Acme Inc</span>
-        </Link>
+        </a>
         <nav className="hidden md:flex gap-4">
-          <Link className="text-sm font-medium hover:underline underline-offset-4" href="#">
+          <a className="text-sm font-medium hover:underline underline-offset-4" href="#">
             Work
-          </Link>
-          <Link className="text-sm font-medium hover:underline underline-offset-4" href="#">
+          </a>
+          <a className="text-sm font-medium hover:underline underline-offset-4" href="#">
             About
-          </Link>
-          <Link className="text-sm font-medium hover:underline underline-offset-4" href="#">
+          </a>
+          <a className="text-sm font-medium hover:underline underline-offset-4" href="#">
             Contact
-          </Link>
+          </a>
         </nav>
       </header>
       <main className="flex-1">
@@ -38,12 +37,12 @@ export function Portfolio() {
                 </p>
               </div>
               <div className="space-x-4">
-                <Link
+                <a
                   className="inline-flex h-9 items-center justify-center rounded-md bg-gray-900 px-8 text-sm font-medium text-gray-50 shadow transition-colors hover:bg-gray-900/90 focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-gray-950 disabled:pointer-events-none disabled:opacity-50 dark:bg-gray-50 dark:text-gray-900 dark:hover:bg-gray-50/90 dark:focus-visible:ring-gray-300"
                   href="#"
                 >
                   Contact
-                </Link>
+                </a>
               </div>
             </div>
           </div>
@@ -115,12 +114,12 @@ export function Portfolio() {
                 </p>
               </div>
               <div className="space-x-4">
-                <Link
+                <a
                   className="inline-flex h-9 items-center justify-center rounded-md bg-gray-900 px-4 py-2 text-sm font-medium text-gray-50 shadow transition-colors hover:bg-gray-900/90 focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-gray-950 disabled:pointer-events-none disabled:opacity-50 dark:bg-gray-50 dark:text-gray-900 dark:hover:bg-gray-50/90 dark:focus-visible:ring-gray-300"
                   href="#"
                 >
                   Contact
-                </Link>
+                </a>
               </div>
             </div>
           </div>
@@ -128,12 +127,12 @@ export function Portfolio() {
       </main>
       <footer className="flex flex-col gap-2 sm:flex-row py-6 w-full shrink-0 items-center px-4 md:px-6 border-t">
         <div className="mx-auto flex gap-4 sm:gap-6">
-          <Link className="text-xs hover:underline underline-offset-4" href="#">
+          <a className="text-xs hover:underline underline-offset-4" href="#">
             Terms of Service
-          </Link>
-          <Link className="text-xs hover:underline underline-offset-4" href="#">
+          </a>
+          <a className="text-xs hover:underline underline-offset-4" href="#">
             Privacy
-          </Link>
+          </a>
         </div>
         <p className="text-xs text-gray-500 dark:text-gray-400">© 2024 Acme Inc. All rights reserved.</p>
       </footer>
@@ -141,8 +140,11 @@ export function Portfolio() {
   )
 }
 
+interface IconProps {
+  className: string;
+}
 
-function MountainIcon(props) {
+function MountainIcon(props: IconProps) {
   return (
     <svg
       {...props}
